Added ModeSortUsermod, modified the other two usermods to use it.
