+++ usermods/usermod_v2_rotary_encoder_ui/usermod_v2_rotary_encoder_ui.h
@@ -9,8 +9,6 @@
 //
 // v2 usermod that provides a rotary encoder-based UI.
 //
-// This Usermod works best coupled with FourLineDisplayUsermod.
-//
 // This usermod allows you to control:
 // 
 // * Brightness
@@ -20,6 +18,11 @@
 // * Palette
 // 
 // Change between modes by pressing a button.
+//
+// Dependencies
+// * This usermod REQURES the ModeSortUsermod
+// * This Usermod works best coupled with 
+//   FourLineDisplayUsermod.
 //
 
 #ifndef ENCODER_DT_PIN
@@ -46,77 +49,6 @@
 // The last UI state
 #define LAST_UI_STATE 4
 
-// Number of modes at the start of the list to not sort
-#define MODE_SORT_SKIP_COUNT 1
-
-// Pointers the start of the mode names within JSON_mode_names
-char **modes_qstrings = nullptr;
-
-// Array of mode indexes in alphabetical order.
-byte *modes_alpha_indexes = nullptr;
-
-// Pointers the start of the palette names within JSON_palette_names
-char **palettes_qstrings = nullptr;
-
-// Which list is being sorted
-char **listBeingSorted = nullptr;
-
-// Array of palette indexes in alphabetical order.
-byte *palettes_alpha_indexes = nullptr;
-
-/**
- * Modes and palettes are stored as strings that
- * end in a quote character. Compare two of them.
- * We are comparing directly within either
- * JSON_mode_names or JSON_palette_names.
- */
-int re_qstringCmp(const void *ap, const void *bp) {
-    char *a = listBeingSorted[*((byte*)ap)];
-    char *b  = listBeingSorted[*((byte*)bp)];
-    int i = 0;
-    do {
-        char aVal = pgm_read_byte_near(a + i);
-        if (aVal >= 97 && aVal <= 122) {
-          // Lowercase
-          aVal -= 32;
-        }
-        char bVal = pgm_read_byte_near(b + i);
-        if (bVal >= 97 && bVal <= 122) {
-          // Lowercase
-          bVal -= 32;
-        }
-        // Relly we shouldn't ever get to '\0'
-        if (aVal == '"' || bVal == '"' || aVal == '\0' || bVal == '\0') {
-          // We're done. one is a substring of the other
-          // or something happenend and the quote didn't stop us.
-          if (aVal == bVal) {
-            // Same value, probably shouldn't happen
-            // with this dataset
-            return 0;
-          }
-          else if (aVal == '"' || aVal == '\0' ) {
-            return -1;
-          }
-          else {
-            return 1;
-          }
-        }
-        if (aVal == bVal) {
-          // Same characters. Move to the next.
-          i++;
-          continue;
-        }
-        // We're done
-        if (aVal < bVal) {
-          return -1;
-        }
-        else {
-          return 1;
-        }
-    } while (true);
-    // We shouldn't get here.
-    return 0;
-}
 
 class RotaryEncoderUIUsermod : public Usermod {
 private:
@@ -131,10 +63,14 @@ private:
   unsigned char prev_button_state = HIGH;
   
 #ifdef USERMOD_FOUR_LINE_DISLAY
-  FourLineDisplayUsermod* display;
+  FourLineDisplayUsermod *display;
 #else
   void* display = nullptr;
 #endif
+
+  byte *modes_alpha_indexes = nullptr;
+  byte *palettes_alpha_indexes = nullptr;
+
   unsigned char Enc_A;
   unsigned char Enc_B;
   unsigned char Enc_A_prev = 0;
@@ -150,17 +86,15 @@ public:
      */
   void setup()
   {
-    /**
-     * Sort the modes and palettes on startup
-     * as they are guarantted to change.
-     */
-    sortModesAndPalettes();
-
     pinMode(pinA, INPUT_PULLUP);
     pinMode(pinB, INPUT_PULLUP);
     pinMode(pinC, INPUT_PULLUP);
     currentTime = millis();
     loopTime = currentTime;
+
+    ModeSortUsermod *modeSortUsermod = (ModeSortUsermod*) usermods.lookup(USERMOD_ID_MODE_SORT);
+    modes_alpha_indexes = modeSortUsermod->getModesAlphaIndexes();
+    palettes_alpha_indexes = modeSortUsermod->getPalettesAlphaIndexes();
 
 #ifdef USERMOD_FOUR_LINE_DISLAY    
     // This Usermod uses FourLineDisplayUsermod for the best experience.
@@ -172,37 +106,6 @@ public:
     }
 #endif
   }
-
-/**
- * Sort the modes and palettes to the index arrays
- * modes_alpha_indexes and palettes_alpha_indexes.
- */
-void sortModesAndPalettes() {
-    modes_qstrings = re_findModeStrings(JSON_mode_names, strip.getModeCount());    
-    modes_alpha_indexes = re_initIndexArray(strip.getModeCount());
-    re_sortModes(modes_qstrings, modes_alpha_indexes, strip.getModeCount(), MODE_SORT_SKIP_COUNT);
-    free(modes_qstrings);
-    modes_qstrings = nullptr;
-
-    palettes_qstrings = re_findModeStrings(JSON_palette_names, strip.getPaletteCount());
-    palettes_alpha_indexes = re_initIndexArray(strip.getPaletteCount());
-
-    int skipPaletteCount = 1;
-    while (true) {
-      // How many palette names start with '*' and should not be sorted?
-      // (Also skipping the first one, 'Default').
-      if (pgm_read_byte_near(palettes_qstrings[skipPaletteCount]) == '*') {
-        skipPaletteCount++;
-      }
-      else {
-        break;  
-      }
-    }
-
-    re_sortModes(palettes_qstrings, palettes_alpha_indexes, strip.getPaletteCount(), skipPaletteCount);
-    free(palettes_qstrings);
-    palettes_qstrings = nullptr;
-}
 
   /*
      * connected() is called every time the WiFi is (re)connected
@@ -449,113 +352,4 @@ void sortModesAndPalettes() {
     effectPalette = palettes_alpha_indexes[effectPaletteIndex];
     lampUdated();
   }
-
-  byte *re_initIndexArray(int numModes) {
-    byte *indexes = (byte *)malloc(sizeof (byte) * numModes);  
-    for (byte i = 0; i < numModes; i++) {
-      indexes[i] = i;
-    }
-    return indexes;
-  }
-
-  /**
-   * Return an array of mode or palette names from the JSON string.
-   * They don't end in '\0', they end in '"'. 
-   */
-  char **re_findModeStrings(const char json[], int numModes) {
-    char **modeStrings = (char **)malloc(sizeof (char *) * numModes);
-    uint8_t modeIndex = 0;
-    bool insideQuotes = false;
-    // advance past the mark for markLineNum that may exist.
-    char singleJsonSymbol;
-
-    // Find the mode name in JSON
-    bool complete = false;
-    for (size_t i = 0; i < strlen_P(json); i++) {
-      singleJsonSymbol = pgm_read_byte_near(json + i);
-      switch (singleJsonSymbol) {
-        case '"':
-          insideQuotes = !insideQuotes;
-          if (insideQuotes) {
-            // We have a new mode or palette
-            modeStrings[modeIndex] = (char*)(json + i + 1);
-          }
-          break;
-        case '[':
-          break;
-        case ']':
-          complete = true;
-          break;
-        case ',':
-          modeIndex++;
-        default:
-          if (!insideQuotes) {
-            break;
-          }
-      }
-      if (complete) {
-        break;
-      }
-    }
-    return modeStrings;
-  }
-
-  /**
-   * Sort either the modes or the palettes using quicksort.
-   */
-  void re_sortModes(char **modeNames, byte *indexes, int count, int numSkip) {
-    listBeingSorted = modeNames;
-    qsort(indexes + numSkip, count - numSkip, sizeof(byte), re_qstringCmp);
-    listBeingSorted = nullptr;
-  }
-
-  /*
-     * addToJsonInfo() can be used to add custom entries to the /json/info part of the JSON API.
-     * Creating an "u" object allows you to add custom key/value pairs to the Info section of the WLED web UI.
-     * Below it is shown how this could be used for e.g. a light sensor
-     */
-  /*
-    void addToJsonInfo(JsonObject& root)
-    {
-      int reading = 20;
-      //this code adds "u":{"Light":[20," lux"]} to the info object
-      JsonObject user = root["u"];
-      if (user.isNull()) user = root.createNestedObject("u");
-
-      JsonArray lightArr = user.createNestedArray("Light"); //name
-      lightArr.add(reading); //value
-      lightArr.add(" lux"); //unit
-    }
-    */
-
-  /*
-     * addToJsonState() can be used to add custom entries to the /json/state part of the JSON API (state object).
-     * Values in the state object may be modified by connected clients
-     */
-  void addToJsonState(JsonObject &root)
-  {
-    //root["user0"] = userVar0;
-  }
-
-  /*
-     * readFromJsonState() can be used to receive data clients send to the /json/state part of the JSON API (state object).
-     * Values in the state object may be modified by connected clients
-     */
-  void readFromJsonState(JsonObject &root)
-  {
-    userVar0 = root["user0"] | userVar0; //if "user0" key exists in JSON, update, else keep old value
-    //if (root["bri"] == 255) Serial.println(F("Don't burn down your garage!"));
-  }
-
-  /*
-     * getId() allows you to optionally give your V2 usermod an unique ID (please define it in const.h!).
-     * This could be used in the future for the system to determine whether your usermod is installed.
-     */
-  uint16_t getId()
-  {
-    return USERMOD_ID_ROTARY_ENC_UI;
-  }
-
-  //More methods can be added in the future, this example will then be extended.
-  //Your usermod will remain compatible as it does not need to implement all methods from the Usermod base class!
 };
